Add "webrick" to Gemfile (#2120)
As of Ruby 3.0 webrick is no longer part of the Ruby standard library:
https://www.ruby-lang.org/en/news/2020/12/25/ruby-3-0-0-released/

Adding it for Ruby 2 in the "test" section should be fine also.

A recent update happened to add webrick to the Gemfile.lock indirectly, when
"yard" was updated since the new version of yard requires webrick. But it's
probably worth adding webrick to the Gemfile explicitly since we depend on it
for tests, not just for yard.