Use field.name rather than resource_name for has_one relationships
If a model has multiple has_one relationships to the same model (e.g. two polymorphic associations with different scopes), we can't override the field label for the two associations, we can only change the translation for both at the same time, as the translation is based on the associated class name, not the association name.

This change uses the association name as the lookup field and default value for the label for has_one relationships, allowing multiple relations to the same class to have their own translated labels.