Document using custom assets
We weren't documenting any custom assets, which is a common thing for
people to want to do. This does that and also provides an example change
in the `example_app`.

Example of custom CSS

Example of custom JS

Document the feature

Avoid scss to avoid sassc error

If using scss here, you'll get `cannot load such file -- sassc`.
Perhaps not if your app is configured to digest scss, but this one
isn't.