Start bundling compiled assets in the Gem
Taking inspiration from how Avo handles their gem assets, this bundles a
compiled version of our assets inside the gem.

We're using `jsbundling-rails` with `esbuild`, along with `cssbundling-rails`
with `saas` to build Administrate's asset files, which is the closest of
the newer approaches for handling assets in Rails as to what we'd been doing
previously.

We then remove `jquery-rails`, `sassc-rails` and `selectize-rails` gems
which are replaced with the equivalents from `npm`.

When releasing, we'll need to run `yarn run build` and `yarn run
build:css` to compile the assets before building the gem which will now
ship the contents of `app/assets/builds` with it.

For someone using Administrate, they'll no longer need to compile
the assets shipped by it as these will now be ready-to-use `.js` and
`.css` files and any asset bundling approaches we choose (or need to,
because they've changed upstream in Rails) should no longer have an
impact on users of the gem.

This is part of: #2311

https://avohq.io/blog/how-to-bundle-assets-in-a-rails-engine

Co-authored-by: Elias Rodrigues <eliasrodrigues.info@gmail.com>
Co-authored-by: Pablo Brasero <pablo@pablobm.com>