Switch to using @thoughtbot/stylelint-config
In #2397, we're having trouble with ignoring directories using the
existing `scss-lint` configuration on Hound. `scss-lint` has been
deprecated for a while and we have another configuration based on
thoughtbot guides, so this is a good fit.

We use BEM style selectors in our CSS. So we add a custom rule, based on
some prior work by @a-tokyo, but adjusted to fit in our JSON
configuration file.

In addition, we disable these rules as they don't make sense on
Administrate:

* Disable scss/dollar-variable-empty-line-before on variables
* Disable function-no-unknown/no-global function names

https://github.com/thoughtbot/stylelint-config
https://github.com/sds/scss-lint/commit/4ac76f9575464028447cbc4042d92eb1089d3899
https://github.com/humanmade/coding-standards/issues/193#issuecomment-1405099508
https://github.com/simonsmith/stylelint-selector-bem-pattern/issues/23