Lint Ruby with standardrb
Some time ago, we at thoughtbot stopped using our custom Rubocop rules
and replaced it with standard. This adds Standard, sets up support for
it in GitHub Actions and runs it against the code base.

We ignore `Lint/ConstantDefinitionInBlock` in a few specs, as this is
related to the Pundit integration. It'd be good to come back to this in
future and adjust the specs to avoid defining/overriding classes in
specs, but honestly it was too difficult to make it work in a reasonable
time.

Finally, we need to use `main` of the standard Action as, currently,
that's the only one which is working (`0.0.5` fails silently).

https://github.com/standardrb/standard
https://github.com/thoughtbot/guides/pull/606