Fix "search with a filter with arguments" spec
The flake would happen when the second search of `kind:vip` was not
applied to the URL. Running the test directly in Chrome would show it
happening. As the URL parameter was never set, it was therefore blank.
Whilst a `sleep(0.1)` would fix it, this instead restructures the test
to avoid the problem completely.

Closes #2523