Rephrase the Administrate pitch
In #2505, we stopped saying that Administrate doesn't have a DSL, but
that pointed to some other things which are outdated in the README and
Gemspec.

This goes further with the rephrasing to remove some weasle words,
comparisons to other frameworks (which made more sense when it was first
released) and try to improve the vision we're pitching for Administrate
now and into the future.