Add tooltip to search
You can add collection filters to a dashboard and then use them in the
search bar on the index page. eg. `email:jared@example.com`

However, these collection filters are not discoverable. You have to know
about them already to know how to use them.

This change introduces a tooltip which lists the available collection
filters, allowing all users to discover what is available.

The tooltip icon is `question-mark-circle` from [heroicons] (MIT licensed)

We use [anchor positioning] with a [polyfill] to position the [popover] by
the tooltip.

Some basic JavaScript is introduced to show/hide the popover on hover to
ensure the tooltip behaviour works as expected for a tooltip.

I've had a first pass at the design, borrowing from existing styles. But
I don't feel strongly about it if we want to go in a different direction.

Do we want an option to disable the tooltip even if there are collection
filters?

[heroicons]: https://heroicons.com/solid
[anchor positioning]: https://developer.mozilla.org/en-US/docs/Web/CSS/position-anchor
[polyfill]: https://github.com/oddbird/css-anchor-positioning
[popover]: https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/popover

Closes https://github.com/thoughtbot/administrate/issues/2750