Introduce acceptance testing
This is the culmination of several months work, and at least one side
quest, in laying down the foundations of a test suite which will run
Administrate completely from scratch. For each scenario, we create a new
Rails application in a clean environment, install Administrate and run
the generators. We can then test against this fresh application.

Whilst we seem to have had a good track record when it comes to past and
future compatibility, the hope is that by having such an outside
perspective focused test suite, this will provide much broader
confidence, especially where different frontend configurations are
concerned. We'll also be able to test speculatively, such as against the
`main` branch of Rails which isn't something we could've done with
confidence in the past with so many different components involved.

To do this, we introduce the `jet_black` gem, which helps with running
"black box" tests for command line tools, and `sack_race`, which helps
run and manage processes inside the test suite.

To run the tests themselves, we use Capybara like we do elsewhere, only
pointed against the Rails project we created as part of the test. It's
not expected that we'd re-use our existing Capybara tests here, but
rather create new ones for each scenario.

As these tests are always going to be slow, we don't run these as usual
when running `rspec`, so they excluded. We'll run them in GitHub Actions
as standalone jobs.

Over time, it's expected that we'd create acceptance tests for all
supported configurations, covering different frontend tools, to projects
with complex model configurations and beyond.

https://github.com/odlp/jet_black
https://github.com/nickcharlton/sack_race