Fix datetime formatting and timezone handling
This ensures consistency across public methods for `DateTime`, `Date`,
and `Time` which all had slight differences.

We remove the `default` option from `I18n.localize`, because an
exception is thrown with `data.in_time_zone` when the `data` is `nil`,
the `default` option is not meaningful.

We don't need a fallback for the time zone, as with `Time.zone`, one is
already defined. We also don't need to do a string conversion, as
`in_time_zone` supports accepting a `Time.zone` object.